don't redownload pages that already exist

+++ validate.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 for x in *.php; do
-    wget -c albanyrowingcenter.org/$x -O /tmp/$x.html;
+    [ -e /tmp/$x.html ] || wget -c albanyrowingcenter.org/$x -O /tmp/$x.html;
 done
 for x in /tmp/*php.html;do
     validate  $x;
